sentinel-nikola: update to version 3.0.0 and rework setup
This reworks how we trigger logging of some firewall zone. We now
provide management script for it. The idea of it is that user can
disable default wan logging and it would not be re-enabled with package
update as well as he can add additional interfaces.
Setting is stored in dedicated uci config to firewall zone that is used
to detect and potentially in future fix/migrate configuration.