reforis: include changes from 0.9.0 version
Version 0.9.0 is prepared for changes, which are included in Turris OS
5.1. There are used package lists instead of user lists. Package lists
allow us to add more features in Updater tab it means there are labels,
descriptions and so on. This change is not backward compatible.

However, there is a possibility to release a new reForis release
(e.g reForis 0.8.1) or choose this approach. It is basically the same.