tos3to4: I found out that I forgot that Turris 1.x has different WAN
The difference and why I missed that in testing was that I was using
port six on switch to CPU and because WAN port was moved to correct
output. The problem is that this is not always true and if users uses
eth1 with switch port 0 then it ends up with invalid config.

This just checks if model is Turris 1.x (on Turris OS 3.x marked just as
Turris) and replaces constant for Omnia with variables.