sentinel-firewall: add firewall configuration loader for sentinel
This introduces new approach to configure updater. The problem with
previous solution, that was in the fact grand new, was it overall
chatiness. On one side it allowed users to modify and create their own
firewall. On the other side it was easy to break it and required lot of
things to go right.

Few notable problems are:
* Updates of user specified rules if configuration needed for Sentinel
  to work changes is pretty much impossible.
* Firewall3 in reality ignores in some cases (as for example mangle)
  source and target zone specification.
* Not all configurations supported by firewall3 are visible in Luci and
  modifying such rule in Luci resulted in wrong config. This was most
  probably caused by invalid rule interpretation in Luci and could be
  fixed.
* Every projects had their own rule to bypass dynamic firewall and
  duplicate co to add it to firewall configuration.

Replacing UCI configured rules with script that inserts appropriate
rules gives us more freedom in how we do stuff. It is for example easy
to more precisely filter traffic. It is possible to insert rules to the
beginning to prepend all other possible rules (that was not as easy done
with just using UCI configs).

In core this package only adds import rule to UCI firewall for script
that later runs all executable files from specific directory. This
script based design also easily allows sharing of the code and
functionality. This means for example that dynamic firewall knows how to
bypass its blocking rule and provides library to other scripts to just
specify port and protocol to bypass.