tos3to4: add SSH daemon config migration
In Turris OS 3.x there was UCI config and integration for sshd. Now
there is just standard config. Simple fix is to move original generated
one to appropriate location.

This move only happens if configuration of sshd in UCI is not the
default one. The second check is if target file exists and that is just
to make sure but it should always be there.

Thanks to David Hopfmueller for pointing this out.