sentinel-firewall: reorder firewall config on start
This checks and reorders sentinel-firewall config on firewall start.
This is not ideal as this reorder is not applied in firewall run when
reorder happens and there seems to be no reliable way to request
firewall reload from firewall script itself (would cause firewall
nesting with unknown results).

We also does not have to reorder in uci-defaults as new config sections
are always appended so they are the last one, making that operation
useless code.