updater-ng: implement target branch bootstrap
This changes how bootstrap is performed. Previously configuration file
was generated by script in turris-build repository when medkit was
generated. This generalizes this approach and instead of generating it
somewhere else we define common BOOTSTRAP variables optionally defined
in environment when bootstrapping root filesystem.
We do not immediately apply them but we rather just store value and set
it on first boot (using uci-defaults). The reason is so we can use uci
instead of directly modifying configuration files.