l10n-supported: implement configuration bootstrap
This changes default bootstrap that was always setting just two (and
current Foris language) languages (that is cs and de). Now it reads same
environment variable as updater's Lua scripts that include those
language packages to system.
We do not immediately apply it to configuration but we rather just store
value and set it on first boot (using uci-defaults). The reason is so we
can use uci instead of directly modifying configuration files.